% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{citation_list_make}
\alias{citation_list_make}
\title{Create crossref citation_list}
\usage{
citation_list_make(ref_path = here::here("data-raw", "metadata",
  "references.csv"))
}
\arguments{
\item{ref_path}{path to \code{references.csv} (defaults to \code{data-raw/metadata/references.csv})}
}
\value{
a crossref \code{citation_list} in list format. To be supplied to the \code{metadata} element of \code{additionalMetadata}.
}
\description{
Create a citation_list of unstructured citations following crossref schema from \code{references.csv}.
}
