% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{classify_validname}
\alias{classify_validname}
\title{Classify worrms validated names}
\usage{
classify_validname(valid_name, ranks = c("phylum", "class", "order",
  "family", "genus", "species"))
}
\arguments{
\item{valid_name}{a worrms validated name}

\item{ranks}{ranks of interest to be returned}
}
\value{
data.frame of ranks for worrms valid name
}
\description{
Classify worrms validated names
}
