% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{ref_add_manual}
\alias{ref_add_manual}
\title{Add manually corrected references}
\usage{
ref_add_manual(references, ref_manual)
}
\arguments{
\item{references}{references data.frame}

\item{ref_manual}{manual references to add}
}
\value{
ref_manual  with references added
}
\description{
Add manually corrected references
}
