% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{validate_taxonomy}
\alias{validate_taxonomy}
\title{Validate taxonomy}
\usage{
validate_taxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{taxonomy data.frame}
}
\value{
if validated, the taxonomy data.frame. Otherwise an error report
}
\description{
Validate taxonomy
}
